pyproject.toml
[tool.poetry]
name = "iris_project"
version = "0.1.0"
description = ""
authors = ["Groblin <lebedivanm@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.22.4"
torch = "^1.13.1"
pandas = "^1.4.2"
scikit-learn = "^1.3.1"
dvc = {extras = ["gdrive"], version = "^3.30.3"}
hydra-core = "^1.3.2"
lightning = "^2022.10.25"
mlflow = "^2.9.2"
onnx = "^1.15.0"
onnxruntime = "^1.16.3"

[tool.poetry.dev-dependencies]
pre-commit = "^2.21.0"

[tool.black]
line-length = 90
target-version = ["py38"]

[tool.isort]
src_paths = ["ds_project"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.nbqa.config]
black = "pyproject.toml"
isort = "pyproject.toml"
flake8 = "setup.cfg"

[tool.nbqa.addopts]
flake8 = ["--extend-ignore=E402"]

[tool.nbqa.mutate]
black = 1
isort = 1

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
